separate toggle code from application, remove beta_application
this separates all toggling related code so that it is more
self contained and better for consumption by the phonegap app
which makes use of it.

during this refactoring I noticed that application.js is 
actually the same as beta_application.js so I have removed
the latter to make the project easier to maintain.