Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Very true. Thanks for pointing this out. I feel embarrassed I submitted this patch! :-)

I've spent an hour trying to find a suitable way to feature detect this but alas I think for the time being browser sniffing is the best way to go.. :-(

Since AppleWebKit 534 seems to fix it. I would suggest the test should be:

	return agent.match( /AppleWebKit\/53[456789]|AppleWebKit\/5[456789]|AppleWebKit\/[6789]/i ) ? true : false;

this at least targets ICS as well as MobileSafari on iOS 5.0...

Thoughts?

Patch-set: 1