Update patch set 1
Patch Set 1: There's a problem with this change, please improve

This feels wrong; seems to be changing the high-resolution image without changing the styles which sounds like it will be wrong.

Does the stylesheet need to be fixed instead so that Android 2.1's browser sizes the icon correctly? Maybe adding -webkit-background-size?

Patch-set: 1