Update patch set 4
Patch Set 4:

This seems to have broken the outputted html leaving an unclosed div element. Most noticeably on the sections in the beta.

If you open a page with sections you will notice the styling it broken and the footer has been wrapped in the final div with class section which is incorrect.

Patch-set: 4