prevent infinite search
On Nokia N95 if you blur the search box it triggers an ajax
search. When writing results it then focuses the search and the
search then loses its focus if the cursor is not over the search
box. This in effect causes an infinite number of searches

the focus does not seem to serve much purpose* so removing
(*added in fba5b781)

Change-Id: I0f2097fcd54e3a99f12c5cd44f162487e16805b3