Update patch set 1
Patch Set 1:

Agreed! It would make sense to use this instead
/w/resources/jquery/jquery.js

Only problem is that is not minified so under current situation we would also need to minify this. This is not ideal so maybe we should just wait till js is loaded via ResourceLoader and takes care of all this for us.

Patch-set: 1