Merged previous contact us form with MaxSem's new special page magic using HTMLForm.
Extends HTMLForm to give us some more useful stuff, such as:
* 'placeholder' param for textarea fields
* returning form with elements wrapped in divs rather than in a table

This new HTMLForm stuff is supported for hidden fields, text fields,
and textarea fields.

Unfortunately re-implements a couple of HTMLForm methods verbatirm
due to the fact that we can't use late static bindings (PHP 5.2+
requirement) and the fact that PHP is dumb when it comes to
static properties/methods. Hopefully this won't matter if these
changes can be merged to core.

Change-Id: If15938f10f92cea6782d9ab090315a959cb4caed

* Removed SpecialMobileFrontend in favor of MaxSem's multiple
SpecialPage approach.
* Removed LeaveFeedbackTemplate in favor of using SpecialPage
to handle feedback form

Change-Id: I53af375d4d46079cd1db62be8f6504815088dc75