fix for bug 32773
I troubleshooted this bug with lucb1e and discovered that on a
nokia e75 mousedown on the body tag fires continously as a user
scrolls. This causes lots of lag so removing in favour of a blur
event.

note the blur has a timeout so that a user can click on search
results without hiding them

Change-Id: Icbf8e68b1e95f28b318a2b2ec39e5e6aebe59c27