Update patch set 10
Patch Set 10: Looks good to me, but someone else must approve

Patrick explained that the HTML is cached in Varnish, so the token can't be output there. There's a check in the special page that allows UAs that are known not to support JS to pass no token.

This looks good to me overall but I'm a bit concerned that we might have UAs that don't support JS but that we don't know about, or UAs without any JS support at all, and the disable images feature would be broken for the.

Patch-set: 10