Update patch set 10
Patch Set 10: Verified; Looks good to me, approved

To Roan's point, we don't really know what percentage of our users have JS enabled or not supported. It sounds like Patrick and Jon are dreaming up a way for us to start tracking that information. Considering the vast improvement of this changeset over what we had before plus the fact that we haven't been able to dream up viable alternatives, I think that this is fine to get into production for now until we can get an accurate sense of JS support amongst our mobile users.

Patch-set: 10