Update patch set 10
Patch Set 10:

I'm not happy with the change to toggle.js The introduced code should be put in a separate javascript file. As far as I can see this has nothing to do with toggling.

The app wanted to use this toggle code but didn't as it was already doing more than it should and this has made the situation worse.

Patch-set: 10