Update patch set 2
Patch Set 2:

I'm particularly curious about people's thoughts on specials/MobileFeedback.php line 31. I put in a hook to allow the opportunity for someone to configure $wgMFFeedbackLinks in some kind of complicated, potentially expensive way ONLY when getFeedbackHtml() runs. I'm not sure this is the best approach, but I wanted a simple way to be able to do more variable/fancy configurations like what we'll need to do on the WMF cluster when this gets deployed (potential configuration that we'll use for this can be seen here: http://pastebin.mozilla.org/1599582).

Patch-set: 2