Update patch set 1
Patch Set 1: There's a problem with this change, please improve

(1 inline comment)

See inline comment, also there are two typical locations where the links are inserted: lede or infobox (this solution works) and the last section of the article (this solution doesn't work). Looks like we'll have to move some elements to the top of the page in order for this to work reliably.

Patch-set: 1