Update patch set 1
Patch Set 1:

I see that as a separate bug - that we don't allow trailing text after the last section. This is a much larger problem (for instance I could invent another element with absolute positioning and put it in the last section)

Is there a bug open for that?

Patch-set: 1