Update patch set 5
Patch Set 5: There's a problem with this change, please improve

(1 inline comment)

This is certainly up for debate, but I'd way rather see HTML generation happen inside of HEREDOCs rather than with a mishmash of closing/opening PHP tags and inline PHP. While the latter is more typical in a lot of MW, I do not think it is A Good Thing. It is /very/ difficult to read/follow and therefore very difficult to review.

Patch-set: 5