Update patch set 4
Patch Set 4: (1 inline comment)

Adding one inline comment.

Only thoughts are with regards to the rename. 
There seems to be a naming convention of prefixing with ext. followed by extension name (although not sure if that's written anywhere). I think the name is  important for debugging purposes and worry it is less obvious when called mobile which extension this relates to

Patch-set: 4