Update patch set 16
Patch Set 16: There's a problem with this change, please improve

Toggling options on/off is a little funky - I think it's because the area responsive to the toggle tap is very tiny. At first, I thought toggle was completely broken on ICS (4.0.2) but after lots of poking, I realized if I am very precise with my tap the toggle works. Android 2.3.6 seems way more forgiving as long as I tap in the left half of the icon.

Patch-set: 16