Deal with javascript that fails to load
on occasions as bug 37261 reveals strange things can happen
due to caching for example a search bar that is expected to be present
is not causing the beta_opensearch.js module to fail to load. This results
in other modules not loading.

wrapping in a try block to deal with these situations

Change-Id: Icf5c17ba1ed05ed8f9711c63d0d8cc4d8161103a