Update patch set 1
Patch Set 1:

> in that event the links will look like "Regarding me, a person, or a company I represent ()", which I do not think is good user experience

No, the brackets are added by mobile-frontend-leave-feedback-article-name message. If returnto is empty, there will be no brackets at all ($article will be ""). See line 60.

It would need to be done on the links instead of the linkText, though. Thanks for pointing out.
It was quite hidden with the variable variables.

The way you're using a hook to communicate through a global is evil, too.

I'd refactor all that function to reduce the WMF-specific code to just the email names.
What about a getTarget() method which returns either a url or an email address?

Oh, and the $articleFactualLink en url should be protocol-relative. https://gerrit.wikimedia.org/r/10537

Patch-set: 1