make sections visible as soon as toggling code has been initialised (bug 37619)
there have been various reports of the toggling code running and allowing sections
to be expanded but not read. To remedy this problem lets reset the opacity as soon
as the toggling code is initialised. If an exception is thrown or there is a script
loading problem after the toggling code has run but before all scripts have finished
running the sections will be readable

Change-Id: I175f962b35285637ccade4f3c061ca7caf7c85bc