write mobiletoken cookie on the current domain
Currently the mobiletoken is only being written to the domain
specified in useFormatCookieDomain - this is not usually the same
(for instance on en.m.wikipedia.org useFormatCookieDomain is
m.wikipedia.org). Also path is unnecessary as '/' is the default.

this will prevent **a lot** of ajax requests to api.php

also dropped saveMobileToken to help with readability
also made it so the token is saved regardless of the presence of

Change-Id: Icf145c784e41d77edece5b4a62559478f08a0518