Update patch set 1
Patch Set 1:

I may be wrong but it seems like an admin cannot turn this off?

Certain wiki's may have MediaWiki:Common.css pages that are not written for mobile and may break their mobile site with introducing this change. It would be good if an admin encountering this problem could update LocalSettings.php to turn it off.

Patch-set: 1