Update patch set 1
Patch Set 1:

@MaxSem: So, that means this should only be loaded if ResourceLoader is available (i.e. not in the classic mode). Braking them apart should in theory not be an issue, but in practice users assume they are loaded in pair and that assumption is justified imho, and should be held up.

Also, when RL is integrated properly (when is that happening?) this will be taken care of automatically since OutputPage enqueues them through the 'site' module. The core 'site' module contains Common.css/.js and <skinname>.css/.js (which in the case of mobile will be "Mobile" or whatever the skin ID is) - so this change seems redundant?

Patch-set: 1