Update patch set 1
Patch Set 1:

@MaxSem: I'm not misunderstanding it, I understand it very well. But whether what I referred to as "classic" is currently default doesn't matter. My point stands: Please don't decouple common.js/common.css.

If MobileFrontend can't handle site modules yet, that's fine and completely understandable. Common.js often contains stuff that doesn't consider mobile at all, so it shouldn't be loaded for all of mobile. Agreed.

But, it's been this way for a while, so it can probably wait until after RL is implemented. No need to implement a broken solution in the meanwhile.

Added some thoughts to that etherpad.

Patch-set: 1