Update patch set 1
Patch Set 1:

Krinkle I hear your point, but I do not think this qualifies as a 'broken' solution - perhaps an incomplete one. If the lack of common.js support is visibly and well documented, I do not think it's that big of a deal for now nor should it be considered a blocker.

Patch-set: 1