Update patch set 1
Patch Set 1:

Hi Krinkle
Apologies if I've misunderstood your comment
Are you saying that you have '.collapsed { display: none; }' in your Common.css ?

I would hope Common.js adds the collapsed class to the table as otherwise this content would be unreadable to anyone without javascript - in which case I'm not sure what destruction would be caused by simply including Common.css and not Common.js

If any CSS rules are dependent on javascript to render content then this is a bug in the content and should be fixed. This change to mobile will hopefully surface those sorts of issues.

Patch-set: 1