remove 'browser score code'
originally this was added so that modules could prevent themselves
being initialised for certain browsers in a generic way but it was
a silly idea. Modules should resort to their own browser/feature
detection where necessary.

Originally added to avoid search working on opera mini 4 but search
doesn't work on opera mini in general now (we serve a javascript disabled
site) so this is just cruft

Change-Id: I3cd8d077bd3283ced46a93ad31692dfc7aa4dc21