Create change
Patch-set: 1
Change-id: Ie34b117d9baa24056d65551d2556466426d42b14
Subject: Refactored some of the code to add CSS/JS to a mobile page. OutputPage has these great methods that use ->mModules to load what's being added. From what I understood, a lot of that _can not_ be loaded in mobile view. The solution was to build an array for mobile, and use only that, making it impossible for other extensions to add resources to a mobile template. This commit will rather first cleanup ->mModules and than do use the OutputPage methods to add CSS/JS to a page (combined with the mobile-only resources). This way, it is now possible to use the BeforePageDisplayMobile hook to actually have another extension add resources to a mobile page ;)
Branch: refs/heads/master
Commit: 3b7d489ecb0073d04823fbef2ef9e81482eb180a