Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Refactored some of the code to add CSS/JS to a mobile page. OutputPage has these great methods that use ->mModules to load what's being added. From what I understood, a lot of that _can not_ be loaded in mobile view. The solution was to build an array for mobile, and use only that, making it impossible for other extensions to add resources to a mobile template. This commit will rather first cleanup ->mModules and than do use the OutputPage methods to add CSS/JS to a page (combined with the mobile-only resources). This way, it is now possible to use the BeforePageDisplayMobile hook to actually have another extension add resources to a mobile page ;)
Commit: e370b292ff663ac3531c1cb278a8d1228e6efa8f