Update patch set 1
Patch Set 1: There's a problem with this change, please improve

(6 inline comments)

Cheers for this Siebrand. A few comments.
In future i'd recommend splitting these sorts of changes into at least 4 separate patches:
* php cleanup
* javascript cleanup
* css cleanup
* test cleanup

This will help get it merged quicker - took me a while to review everything! Most of my comments are related to javascript so if this was done as a separate patch the css and php could have got merged separately.

Patch-set: 1