Update patch set 2
Patch Set 2: There's a problem with this change, please improve

This appears to address the issues in the first patchset as far as i can tell. However, if you click on an image link, it takes you to the file page for the image, but does not display the image - instead it displays the image link again. And at least on my local test setup, if I click the link on the filepage, it just reloads the file page, which I assume is not intended functionality? I am testing with an export of the SF article and commons images enabled.

Patch-set: 2