Update patch set 1
Patch Set 1:

A few comments..
On a small article in latest beta without this change:
16 requests  ❘  315.99KB transferred

With this change:
18 requests  ❘  408.26KB transferred

San Francisco article
without: 22 requests  ❘  1.03MB transferred
with: 25 requests  ❘  1.09MB transferred

I'm guessing that a lot of our code can be refactored but it would be good to look at what we're getting and whether it is useful on mobile.

I'd be more comfortable if this was beta only so we can assess any potential damage.

Patch-set: 1