Update patch set 1
Patch Set 1: There's a problem with this change, please improve

(1 inline comment)

Currently swiping when you are down the page makes the user lose their place in the article. It should either only work when at the top of the page or the left menu should be fixed.

(FWIW as it currently stands I don't think there is much use in being able to slide open the left menu - I don't think the existing functions in this menu warrant constant accessibility in comparison to things like the watch icon). However if the watch icon was in the left menu or a similar right menu that would be a totally different story..

Patch-set: 1