(bug 41509) Fix styling of galleries on mobile
The stylesheet still assumed table.gallery, but we switched to a div
based solution quite a while ago. This is a straight copy of the style
from common/shared.css of core with some less flavours

Add !important overrides for the widths of galleries.
original patchset by thedj adjusted by Jon Robson

Change-Id: Ib3a6e258c60ac5f4083dfe72e2e9e834dfa3d112