Update patch set 2
Patch Set 2: Verified; Looks good to me, approved

This particular changeset looks ok, but the formatting of the message seems off to me - Jon, I'll send you a screenshot. Merging as the formatting looks busted w/o this changeset but wanted to put it on your radar.

Patch-set: 2
Label: SUBM=+1