Update patch set 1
Patch Set 1:

The if statement you've suggested does not do the same thing and would not address the original bug.

The list of languages contains the current language. There seems to be an edge case in the case of simple wiki where there are two references to the language en, one of which is simple wiki, the other English.

In the current setup because of this we show that there are 2 languages available but don't print them as those 2 languages are the current language - so production looks strange. I'd rather apply this fix as the problem is very visible and then address the simplewiki problem separately later.

Patch-set: 1