Update patch set 4
Patch Set 4: (4 inline comments)

I've added a few more comments.
it's not really clear what problem these rules are supporting and if they are in fact needed.
I'd be tempted to remove them altogether and fix on a case by case basis if we are not clear what all these rules exist for.

Patch-set: 4