Update patch set 1
Patch Set 1:

Corresponding configuration change: https://gerrit.wikimedia.org/r/#/c/34183/

Max I thought about doing that but decided to just remove the code to reduce complexity - some of the stuff in here is already too-configurable (mixing a few different attempts at providing the contact form the way product wanted it) and the easiest solution for now seemed to be to just get rid of it. We can always add it back IF we need to. If you feel strongly about it though, I don't mind reworking this to be a configurable change.

Patch-set: 1