Update patch set 1
Patch Set 1:

Fun... I'll take a peek at this. Parser cache should already include variant in its key, but if we're doing separate caching that'll need to be tweaked.

Also, I need to add support for converting the page title and section titles so still need to get that variable. :)

Patch-set: 1