Update patch set 1
Patch Set 1:

Nope.. but in my opinion the current experience is completely broken (it's a bug) and we should experiment with alternatives. Since this is beta only and doesn't effect the editing experience this shouldn't be a problem?

Can you demonstrate this to product in my absence to check their okay with it?

Patch-set: 1