Update patch set 3
Patch Set 3:

This commit is part of a larger group of commits that has a significant impact on how tables work in MF - and we agreed to run changes like this by product back at the end of October. My comment wasn't necessarily directed to this particular changeset but rather intended for the group of changes that this is a part of - sorry if that was not clear. But since you've now run this by product and they seem stoked about it, this is fine for merge pending review.

Patch-set: 3