(bug 41514) Hash links broken in mobile
window.history.replaceState() does not require the third
parameter. When it's missing it just uses current URL. Previously
current URL was generated but without the hash and that's why the hash
got "swallowed."

Moreover, we need to know which section should be expanded when a
hash/anchor refers a subsection. Since subsections in beta are
inserted into DOM dynamically, this information needs to be stored
separately (anchorSection object).

The following
    if ( hash.indexOf( '#' ) === 0 ) {
in checkHash() was always true, so I've removed it.

Change-Id: I2f3b8173321317bc58b213a84e73a674b6f28729