Update patch set 1
Patch Set 1:

I assume you are referring to the commit 
https://gerrit.wikimedia.org/r/#/c/32560/
(not this one)

Yes! This is the expected behaviour. Many people were commenting that they needed more than the title. This was one of the changes in "adjust random function" - "when clicking random load the random article" (as in load the whole article not just the title)

Patch-set: 1