Update patch set 2
Patch Set 2: There's a problem with this change, please improve

There's a problem with performance. If you keep clicking "Try again", after 10-20 times it gets really slow. It seems like it's JS because the browser simply freezes up for a few seconds (or more). My guess is that DOM elements are not removed properly or there's some other kind of memory leak.

I tested it on Chrome, master seems to be working fine (maybe something with loading 10 articles at once?).

Patch-set: 2