Update patch set 7
Patch Set 7: There's a problem with this change, please improve

The edit button shows when logged in and in beta and in an article.
The edit link shows in the top right corner.
With this change the edit link disappears as content_wrapper is not position relative by default and the edit link is position absolute.

You'll need to find some other way of achieving the goal.

I'm not sure we should be using position absolute on mw-mf-userinfo anyway..

Patch-set: 7