Update patch set 16
Patch Set 16:

In response to Brion:

1) still doesn't work in Firefox -- 403 error spews to console, spinner keeps going
- now disabled for Firefox - webkit only.

2) initial view post-save doesn't match defined thumbnail size
I don't think we should block on this but I do agree we should fix this. We should also ensure the uploaded file post save links to the file page. I'd rather get this out into the wild to try out the feature on an audience first though.

3) mysterious error messages in Chrome on Android "Error whilst uploading, try another" -- turned out to be because page was cached and I wasn't logged in. Need better reporting for token or upload authentication errors?

jgonera added better error handling so this should be taken care of. If api returns an error it is printed.

4) clicking 'edit' on page after uploading sends to series of infinite reidercts to an edit page that says "there is currently no text in this page" (Chrome and Firefox)

not introduced by this patchset - looking at this separately.

5) clicking on license link opens it in the same window; when going back to the previous page we may lose data such as the file we've just selected

fixed.. although hacky fix.

Patch-set: 16