Update patch set 1
Patch Set 1: There's a problem with this change, please improve

Since this code is a hackish workaround we'll probably want to get rid of later, I'd put a comment referencing the bug next to the changed lines. If there's a WebKit bug that's causing that, I'd reference it too so that we can track it (or maybe reference the WebKit bug in our bugzilla).

Patch-set: 1