Update patch set 2
Patch Set 2:

It seems a bit harsh to -2 this. I'm not quite clear what discussion/input is needed other than changing the copy text (which is easy enough to do later) and styling it (which looks like it will be covered in https://mingle.corp.wikimedia.org/projects/mobile/cards/265 and I feel the beta is the place to get this in front of the most faces to get that input.

Please at least mark as a -1 with concrete ways forward.
I'd hate to end up writing this code again which I predict I will if this is completely abandoned.

Patch-set: 2