Update patch set 2
Patch Set 2:

I'm still trying to understand the use case - if the idea is to create additional menu items would it not make sense to add additional items to the menu (the one you open when you click the 3 line icon in the top right) [1].

I could imagine a hook that allows you to add new menu items in that list with an icon. For your use case you could then create links to קהילה and Yeshiva.org.il - which seems a better approach (otherwise you are mixing the article content with sections that have nothing to do with the content...

thoughts?

[1] https://gerrit.wikimedia.org/r/gitweb?p=mediawiki/extensions/MobileFrontend.git;a=blob_plain;f=stylesheets/common/images/menu/main.png;h=c0b030a171b0463039915efa75c643e19fde8036;hb=HEAD

Patch-set: 2