Update patch set 3
Patch Set 3: (1 inline comment)

See inline comment.

I'm not married to the name - I called it OverloadTemplate as in the concept of method overloading, however it's a bit of a misnomer. It's not absolutely needed but will help with redundant code. I suppose I could alternatively use the same class for the signup form as the login form, but it felt cleaner to me to keep them separate.

Patch-set: 3