Update patch set 3
Patch Set 3:

This smells a bit like premature optimisation?

We should discuss where the account link goes, but in the meantime I'm happy having it in the menu if this helps us get an account creation feature into the wild quicker - although it should have some sort of icon even if this is just the existing login icon for the time being.

Patch-set: 3