Update patch set 2
Patch Set 2: There's a problem with this change, please improve

Thanks for the patch set! However, I agree with Jon - it sounds like it would make more sense to expose a hook for adding items to the navigation menu, rather than arbitrary content to the end of an article. If you feel you really need to further manipulate the article body content, I think it would make more sense to add a generic hook before the 'bodytext' template var gets set (in SkinMobileBase) to allow for html to be further manipulated (perhaps it could take the template object and the html as params).

Patch-set: 2