Update patch set 1
Patch Set 1: There's a problem with this change, please improve

I'm rather against using <h1> for notofication. If we used <article> and <section> tags in our code, that could be OK, but in this case this seems to give too much importance to the notification. I would change it to <p>.

Patch-set: 1