Update patch set 1
Patch Set 1:

The idea of the designers was to start making the homepage a personalised experience. So really it is not a notification - it is a heading for the page.

Note there is currently no h1 on a homepage and that I am simply following the way the designers asked it to be!

Patch-set: 1