Update patch set 1
Patch Set 1:

We are actually creating the login form in MobileFrontend and don't need to display it inside a table - we can do whatever we want! I just did it in a table because that was how the login form was laid out. Take a look at /includes/skins/UserAccountCreateMobileTemplate.php - we use this to override the create account form output, so we have total control of the form.

Patch-set: 1