Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

This looks better to me now (though I haven't tested it yet). I was originally thinking we should cache results in memcache if possible, but I just looked through Special:Watchlist code and there's no use of memcache there, so if it works as-is in production, we shouldn't need it here either.

Patch-set: 3