Update patch set 1
Patch Set 1: There's a problem with this change, please improve

from preilly's email, it sounds like the redirect stuff may still be needed:
"This was for search redirection. Basically the search and special
pages would redirect the user away from m to the desktop site without
this change in place."

Max can you confirm whether or not we still need this?

Patch-set: 1