Update patch set 1
Patch Set 1: There's a problem with this change, please improve

It's very good that you pointed out a localization issue with this patch. Perhaps, we need a message like MediaWiki:Disambiguationspage, but for category (e.g. MediaWiki:Disambiguationscategory). In this case we can just retrieve it with mw.msg(); I believe MobileFrontend client scripting is delivered by RL, isn't it?

Patch-set: 1