Update patch set 1
Patch Set 1:

Zeljko, my concern that it's not a test repository - it's a MediaWiki extension that stores its tests with itself. So people would expect something related to MobileFrontend features in features/, not some tests 99% of them are unable to understand. Let's not mix code with tests in the same tree.

Patch-set: 1