Update patch set 2
Patch Set 2: There's a problem with this change, please improve

Is there official design for this? That drawer that drops down just to show a message seems awkward (and too big). I have a feeling that we're getting too many different widgets in the UI. The toasts (and I guess we have more than one type of them, one with black and other with grey background) and now the drawer. And they're all serving the same purpose.

Patch-set: 2