Update patch set 2
Patch Set 2:

This is definitely an improvement. The only improvements I'd make is to put a margin-top on the headers (5px should be fine) so the sections are more scannable.

In the future, I think we should do a few extra things and maybe make a Mingle card out of it:

1. Include icons in the headers so they're more easily scannable for the colorblind.

2. Auto-compress the diffs to ensure that they fit on the screen and include an interaction to let the user see the whole thing, if they want.

3. Consider using "words" instead of "bytes" for new users as a default, though this is potentially controversial.

Patch-set: 2