Update patch set 1
Patch Set 1:

Re-adding the hamburger icon is a separate discussion. This was a preliminary step (and yes it was intended to remove it from all special pages but keep on javascript overlays)

This shouldn't be too confusing as clicking the hardware back button on the browser will take you back to where you were before.

Currently the back button is broken on diff pages and is ambiguous - does clicking back take you back to the item you clicked on with the same filter? This is hard yet the hardware back button does exactly this.

Worth grabbing Munaf and Maryana and clarifying this.

Patch-set: 1