Update patch set 4
Patch Set 4: There's a problem with this change, please improve

(3 inline comments)

I'm not saying you should change anything in mf-watchlist.js for this changeset. I'm just making a point. I'll prepare a draft this week of how we could reorganize the code.

The only small reason for -1 is my doubt in mf-settings.js.

Patch-set: 4