Update patch set 1
Patch Set 1: There's a problem with this change, please improve

(1 inline comment)

Looks good, some thoughts:

When the roles div is empty (and I guess it can be), the edits div is very close to the user name, it would be good to add top margin to it.

When it comes to separating edit count from message, maybe ::first-line and fixed width of the edits div could be used (which would make "edit(s)" wrap to the second line). Or maybe we could change the message string to '{{PLURAL:$1|edit|edits}}' (no $1 at the beginning). Although, maybe there are languages where "edits" comes before the number which would mean that we should rethink the design.

Also, the title of the article should be black, sans-serif and closer to the "bytes added".

Patch-set: 1