Update patch set 1
Patch Set 1: (1 inline comment)

The essence of the patch set it great - there's no reason we should be linking to js scripts if the phone doesn't support it. That said, this patch set doesn't go far enough - we shouldn't perform /any/ operations in this method around JS scripts if js is not enabled on the phone.

Patch-set: 1