Update patch set 4
Patch Set 4: There's a problem with this change, please improve

You can't merge this. It contains the biggest i18n type of error possible.

The screenshot is English and LTR. It should display an actually RTL UI in RTL use cases and the text has to be i18n-able, OR I guess blending out the two strings will go a long way, because the screenshot is all about the icon.

Jon and I had a discussion on IRC about this, so he will be able to provide some additional background.

Patch-set: 4