Update patch set 6
Patch Set 6: There's a problem with this change, please improve

I still don't like the length of the showResults() function.

Also, there's many things wrong with this. First of all, it limits the number of items in the "recently changed" view too, but doesn't show a "more" link there.

Secondly, it would be really great if "more" could be styled just as a "Sign in" or "Read in another language buttons" button (agora button) and centered, *not* as a list item, so that everything doesn't start to get ugly when https://gerrit.wikimedia.org/r/#/c/41205/ gets merged.

Patch-set: 6