Update patch set 1
Patch Set 1:

according to Max this is not the "mediawiki way" and I wasn't sure how best to do this. Rather than be blocked on this I would be keen to resolve this issue asap as it causes 500's on the server and worry waiting on a proper implementation for 404s will delay this fix.

Patch-set: 1