Update patch set 7
Patch Set 7:

Max your -1 makes no sense - the desktop view doesn't have a list view which does need it - what good is a list if you can't see the entire list?

The feed view gets it for free and if we see performance issues or any other pitfalls with this then by all means we can remove it before moving this to "stable".

Patch-set: 7