Update patch set 1
Patch Set 1: There's a problem with this change, please improve

That's a great idea.

I'm not sure though if handlebars is the right choice for us. It's pretty heavy for mobile, 40KB uglified. I'd use mustache or hogan (9KB uglified) or possibly the one you found a few days ago (Transparency, 10KB uglified). Let's discuss this when you're back.

Patch-set: 1