Update patch set 5
Patch Set 5:

the patchset i just uploaded should fix the issue. however, i assume that Max introduced the previous patchset because controlling the context from within the skin is kinda inappropriate, which made me wonder if setOverlay() even belongs in MobileContext. we should probably rethink this, but at least this patchset seems to make it go for now :)

Patch-set: 5