(bug 43909) (mingle 377) Include article to watch information in login form
If a user is coming to the login form from tapping the 'add to
watchlist' icon on an article, this will include the article to
be watched in a hidden field in the login form. This will later
be used to actually watch that specific article after login is
complete.

Change-Id: I8b7263557761610ffd87f8793d01775b5c85610b